Finding loops in contact map

Find loops using chromosight

getLoops(
  x,
  resolution = NULL,
  output_prefix = file.path("chromosight", "chromo"),
  norm = "auto",
  max.dist = "auto",
  min.dist = "auto",
  min.separation = "auto",
  n.mads = 5L,
  pearson = "auto",
  nreads = "no",
  ncores = 1L
)

Arguments

x

A HiCExperiment object

resolution

Which resolution to use to search loops

output_prefix

Prefix to chromosight output (default: "chromosight/chromo")

norm

Normalization parameter for chromosight

min.dist, max.dist

Min and max distance to use to filter for significant loops

min.separation

Minimum separation between anchors of potential loops

n.mads

Number of MADs to use to filter relevant bins to search for loops

pearson

Minimum Pearson correlation score to use to filter for significant loops

nreads

Number of reads to subsample to before searching for loops

ncores

Number of cores for chromosight

Value

A HiCExperiment object with a new "loops" topologicalFeatures storing significant interactions identified by chromosight, and an additional chromosight_args metadata entry.

Examples

contacts_yeast <- contacts_yeast()
#> snapshotDate(): 2023-04-03
#> see ?HiContactsData and browseVignettes('HiContactsData') for documentation
#> loading from cache
contacts_yeast <- getLoops(contacts_yeast)
S4Vectors::metadata(contacts_yeast)$chromosight_args
#> $`--pattern`
#> [1] "loops"
#> 
#> $`--dump`
#> [1] "/tmp/RtmpmcH5tq"
#> 
#> $`--inter`
#> [1] FALSE
#> 
#> $`--iterations`
#> [1] "auto"
#> 
#> $`--kernel-config`
#> NULL
#> 
#> $`--perc-zero`
#> [1] "auto"
#> 
#> $`--perc-undetected`
#> [1] "auto"
#> 
#> $`--tsvd`
#> [1] FALSE
#> 
#> $`--win-fmt`
#> [1] "json"
#> 
#> $`--win-size`
#> [1] "auto"
#> 
#> $`--no-plotting`
#> [1] TRUE
#> 
#> $`--smooth-trend`
#> [1] FALSE
#> 
#> $`--norm`
#> [1] "auto"
#> 
#> $`<contact_map>`
#> [1] "/github/home/.cache/R/ExperimentHub/54b22d00668c_7752::/resolutions/16000"
#> 
#> $`--max-dist`
#> [1] "auto"
#> 
#> $`--min-dist`
#> [1] "auto"
#> 
#> $`--min-separation`
#> [1] "auto"
#> 
#> $`--n-mads`
#> [1] 5
#> 
#> $`<prefix>`
#> [1] "chromosight/chromo"
#> 
#> $`--pearson`
#> [1] "auto"
#> 
#> $`--subsample`
#> [1] "no"
#> 
#> $`--threads`
#> [1] 1
#> 
topologicalFeatures(contacts_yeast, 'loops')
#> GInteractions object with 53 interactions and 5 metadata columns:
#>        seqnames1       ranges1     seqnames2       ranges2 |   bin_id1
#>            <Rle>     <IRanges>         <Rle>     <IRanges> | <numeric>
#>    [1]        II   16001-32000 ---        II 416001-432000 |        16
#>    [2]        II   16001-32000 ---        II 624001-640000 |        16
#>    [3]        II 160001-176000 ---        II 336001-352000 |        25
#>    [4]        II 432001-448000 ---        II 576001-592000 |        42
#>    [5]       III   16001-32000 ---       III 176001-192000 |        67
#>    ...       ...           ... ...       ...           ... .       ...
#>   [49]       XVI   16001-32000 ---       XVI 272001-288000 |       704
#>   [50]       XVI 304001-320000 ---       XVI 384001-400000 |       722
#>   [51]       XVI 400001-416000 ---       XVI 672001-688000 |       728
#>   [52]       XVI 720001-736000 ---       XVI 928001-944000 |       748
#>   [53]       XVI 704001-720000 ---       XVI 768001-784000 |       747
#>          bin_id2     score    pvalue    qvalue
#>        <numeric> <numeric> <numeric> <numeric>
#>    [1]        41  0.367443 6.316e-07 6.832e-07
#>    [2]        54  0.420304 7.000e-09 9.600e-09
#>    [3]        36  0.492732 0.000e+00 0.000e+00
#>    [4]        51  0.417455 0.000e+00 0.000e+00
#>    [5]        77  0.458787 2.100e-09 3.100e-09
#>    ...       ...       ...       ...       ...
#>   [49]       720  0.587076 0.000e+00 0.000e+00
#>   [50]       727  0.326277 5.085e-07 5.734e-07
#>   [51]       745  0.601166 0.000e+00 0.000e+00
#>   [52]       761  0.567326 0.000e+00 0.000e+00
#>   [53]       751  0.333732 5.591e-07 6.174e-07
#>   -------
#>   regions: 98 ranges and 0 metadata columns
#>   seqinfo: 14 sequences from an unspecified genome; no seqlengths